airflow-docker/requirements.txt
alembic
apache-airflow
apache-airflow-providers-ftp
apache-airflow-providers-google
apache-airflow-providers-http
apache-airflow-providers-imap
apache-airflow-providers-postgres
apache-airflow-providers-sqlite
apache-airflow-providers-ssh
apispec
argcomplete
astroid
attrs
Babel
bcrypt
cached-property
cachetools
cattrs
certifi
cffi
chardet
click
clickclick
cloudpickle
colorama
colorlog
commonmark
connexion
croniter
cryptography
databricks-cli
dataclasses-json
defusedxml
dill
dnspython
docker
docutils
email-validator
entrypoints
Flask
Flask-AppBuilder
Flask-Babel
Flask-Caching
Flask-JWT-Extended
Flask-Login
Flask-OpenID
Flask-SQLAlchemy
Flask-WTF
gitdb
GitPython
google-ads
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-automl
google-cloud-bigquery
google-cloud-bigquery-datatransfer
google-cloud-bigquery-storage
google-cloud-bigtable
google-cloud-container
google-cloud-core
google-cloud-datacatalog
google-cloud-dataproc
google-cloud-dlp
google-cloud-kms
google-cloud-language
google-cloud-logging
google-cloud-memcache
google-cloud-monitoring
google-cloud-os-login
google-cloud-pubsub
google-cloud-redis
google-cloud-secret-manager
google-cloud-spanner
google-cloud-speech
google-cloud-storage
google-cloud-tasks
google-cloud-texttospeech
google-cloud-translate
google-cloud-videointelligence
google-cloud-vision
google-crc32c
google-resumable-media
googleapis-common-protos
graphviz
greenlet
grpc-google-iam-v1
grpcio
grpcio-gcp
gunicorn
httplib2
idna
importlib-metadata
importlib-resources
inflection
iniconfig
iso8601
isodate
isort
itsdangerous
Jinja2
joblib
json-merge-patch
jsonschema
lazy-object-proxy
libcst
lockfile
Mako
Markdown
MarkupSafe
marshmallow
marshmallow-enum
marshmallow-oneofschema
marshmallow-sqlalchemy
mccabe
mlflow
mypy-extensions
natsort
nltk
numpy
oauthlib
openapi-spec-validator
packaging
pandas
pandas-gbq
paramiko
pendulum
pluggy
prison
prometheus-client
prometheus-flask-exporter
proto-plus
protobuf
psutil
psycopg2-binary
py
pyarrow
pyasn1
pyasn1-modules
pycparser
pydata-google-auth
Pygments
PyJWT
pylint
pymystem3
PyNaCl
pyOpenSSL
pyparsing
pyrsistent
pysftp
pysqlite3
pytest
python-daemon
python-dateutil
python-editor
python-nvd3
python-slugify
python-youtube
python3-openid
pytz
pytzdata
PyYAML
querystring-parser
regex
requests
requests-oauthlib
rich
rsa
scikit-learn
scipy
setproctitle
six
smmap
SQLAlchemy
SQLAlchemy-JSONField
SQLAlchemy-Utils
sqlparse
sshtunnel
stringcase
swagger-ui-bundle
tabulate
tenacity
tensorflow
termcolor
text-unidecode
threadpoolctl
toml
torch
tqdm
typing-extensions
typing-inspect
unicodecsv
uritemplate
urllib3
websocket-client
Werkzeug
wrapt
WTForms
zipp
